update cmssw_l2/check.sh to use cherry-pick instead of checkout

+++ cmssw_l2/check.sh
@@ -1,10 +1,9 @@
-#!/bin/bash -e
+#!/bin/bash -ex
 push_chg=true
 if [ "$1" = "" -o "$1" = "false" ] ; then push_chg=false ; fi
-sdir=$(dirname $0)
+sdir=$(realpath $(dirname $0))
 [ -f ${sdir}/l2.json ] || echo '{}' > ${sdir}/l2.json
-old_commit=""
-[ -f ${sdir}/commit.txt ] && old_commit=$(cat ${sdir}/commit.txt)..
+old_commit="$(cat ${sdir}/commit.txt)"
 
 rm -rf update_cmssw_l2
 git clone -q git@github.com:cms-sw/cms-bot update_cmssw_l2
@@ -12,13 +11,16 @@ pushd update_cmssw_l2
   export PYTHONPATH=$(/bin/pwd -P)
   export PYTHONUNBUFFERED=1
   commit=""
-  for data in $(git log --no-merges --pretty=format:"%H:%at," ${old_commit} -- categories.py | tr ',' '\n' | grep : | tac) ; do
+  git checkout ${old_commit}
+  for data in $(git log --no-merges --pretty=format:"%H:%at," ${old_commit}..master | tr ',' '\n' | grep : | tac) ; do
     commit=$(echo $data | sed 's|:.*||')
     cur_time=$(echo $data | sed 's|.*:||')
-    git checkout -q $commit
-    rm -rf *.pyc __pycache__
-    echo "Working on $commit"
-    ${sdir}/update.py ${sdir}/l2.json ${cur_time} 2>&1
+    git cherry-pick $commit
+    if [ $(git diff --name-only HEAD^ | grep "^categories.py" | wc -l) -gt 0 ] ; then
+      echo "Working on $commit"
+      ${sdir}/update.py ${sdir}/l2.json ${cur_time} 2>&1
+      rm -rf *.pyc __pycache__
+    fi
   done
 popd
 rm -rf update_cmssw_l2
